setup_mysql_db: Added crud ops

+++ src/main/java/com/semanta/share/service/share/ShareService.java
@@ -5,7 +5,7 @@ import com.semanta.share.model.FileInfo;
 import java.util.ArrayList;
 
 public interface ShareService {
-    public abstract String upload(String delOnFirstView, int timeout);
+    public abstract String upload(String delOnFirstView, long timeout);
 
     public abstract ArrayList<FileInfo> retrieve(String dirNonce) throws Exception;
 }
